+++ src/Command/ResetTimestampFreeradiusCommand.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Command;
 
 use App\Entity\Setting;
@@ -14,11 +16,10 @@ use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 #[AsCommand(
-    name: 'reset:ScheduleSettings',
-    description: 'Reset Schedule Settings',
+    name: 'reset:timeStampFreeradiusCron',
+    description: 'Reset Timestamp for Freeradius cron'
 )]
-
-class ResetScheduleSettingsCommand extends Command
+class ResetTimestampFreeradiusCommand extends Command
 {
     public function __construct(
         private readonly EntityManagerInterface $entityManager
@@ -29,17 +30,19 @@ class ResetScheduleSettingsCommand extends Command
     protected function configure(): void
     {
         $this
-            ->setName('reset:ScheduleSettings')
-            ->setDescription('Reset Schedule Settings')
+            ->setName('reset:timeStampFreeradiusCron')
+            ->setDescription('Reset Timestamp for Freeradius cron')
             ->addOption('yes', 'y', InputOption::VALUE_NONE, 'Automatically confirm the reset');
     }
-
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         // Check if the --yes option is provided (comes from a controller), then skip the confirmation prompt
         if (!$input->getOption('yes')) {
             $helper = $this->getHelper('question');
-            $question = new ConfirmationQuestion('This action will reset the Schedule Settings. [y/N] ', false);
+            $question = new ConfirmationQuestion('This action will RESET the TIMESTAMP used by the' .
+                'FREERADIUS cron job on the portal. ' .
+                'This is a sensitive operation and should only be performed in **very specific scenarios**. ' .
+                'Are you sure you want to proceed? [y/N]', false);
             /** @var QuestionHelper $helper */
             if (!$helper->ask($input, $output, $question)) {
                 $output->writeln('Command aborted.');
@@ -48,11 +51,7 @@ class ResetScheduleSettingsCommand extends Command
         }
 
         $settings = [
-            ['name' => 'FREERADIUS_LAST_CONNECTION_CRON', 'value' => '* * * * *'],
-            ['name' => 'DELETE_UNCONFIRMED_USERS_CRON', 'value' => '0 0 * * *'],
-            ['name' => 'USERS_WHEN_PROFILE_EXPIRES_CRON', 'value' => '0 1 * * *'],
-            ['name' => 'LDAP_SYNC_CRON', 'value' => '0 2 * * *'],
-            ['name' => 'CRON_ADVANCED_STATUS', 'value' => 'OFF'],
+            ['name' => 'TIME_STAMP_FREERADIUS_CRON', 'value' => null],
         ];
 
         // Begin a database transaction to ensure data consistency
@@ -85,7 +84,7 @@ class ResetScheduleSettingsCommand extends Command
 
             $message = <<<EOL
 
-<info>Success:</info> All the Schedule Settings have been set to the default values.
+<info>Success:</info> Timestamp for Freeradius cron have been set to the current time.
 <comment>Note:</comment> If you want to reset any another setting please check using this command:
       <fg=blue>php bin/console reset</>
 EOL;