+++ src/Controller/UsersManagementController.php
@@ -2,9 +2,11 @@
 
 namespace App\Controller;
 
+use App\DTO\UserUpdateDTO;
 use App\Entity\Event;
 use App\Entity\User;
 use App\Entity\UserExternalAuth;
+use App\Entity\UserRadiusProfile;
 use App\Enum\AnalyticalEventType;
 use App\Enum\FirewallType;
 use App\Enum\OperationMode;
@@ -16,6 +18,7 @@ use App\Form\ResetPasswordType;
 use App\Form\UserUpdateType;
 use App\Repository\EventRepository;
 use App\Repository\UserExternalAuthRepository;
+use App\Repository\UserRadiusProfileRepository;
 use App\Repository\UserRepository;
 use App\Service\EscapeSpreadSheet;
 use App\Service\EventActions;
@@ -60,7 +63,8 @@ class UsersManagementController extends AbstractController
         private readonly UserDeletionService $userDeletionService,
         private readonly TwoFAService $twoFAService,
         private readonly VerificationCodeEmailGenerator $verificationCodeEmailGenerator,
-        private readonly TranslatorInterface $translator
+        private readonly TranslatorInterface $translator,
+        private readonly UserRadiusProfileRepository $radiusProfileRepository,
     ) {
     }
 
@@ -333,7 +337,6 @@ class UsersManagementController extends AbstractController
     #[IsGranted('ROLE_ADMIN')]
     public function editUsers(
         Request $request,
-        UserRepository $userRepository,
         UserPasswordHasherInterface $passwordHasher,
         EntityManagerInterface $em,
         MailerInterface $mailer,
@@ -363,42 +366,33 @@ class UsersManagementController extends AbstractController
             return $this->redirectToRoute('admin_page');
         }
 
-        // Store the initial bannedAt value before form submission
-        $initialBannedAtValue = $user->getBannedAt();
+        // Prepare DTO
+        $userUpdateDTO = new UserUpdateDTO($user);
+        $initialBannedAt = $user->getBannedAt();
 
-        $form = $this->createForm(UserUpdateType::class, $user);
+        // Create & handle form
+        $form = $this->createForm(UserUpdateType::class, $userUpdateDTO);
         $form->handleRequest($request);
+
         if ($form->isSubmitted() && $form->isValid()) {
-            $user = $form->getData();
+            // Use DTO method to map data back
+            $userUpdateDTO->updateUser($user);
 
-            // Verifies if the isVerified is removed to the logged account
-            if (($currentUser->getId() === $user->getId()) && $form->get('isVerified')->getData() === 0) {
-                $user->isVerified();
-                $this->addFlash(
-                    'error_admin',
-                    $this->translator->trans('administratorsCannotRemoveOwnVerification', [], 'controllers')
-                );
-                return $this->redirectToRoute('admin_user_edit', ['id' => $user->getId()]);
-            }
-
-            // Verifies if the bannedAt was submitted and compares the form value "banned" to the current value
-            if ($form->get('bannedAt')->getData() && $user->getBannedAt() !== $initialBannedAtValue) {
-                if ($currentUser->getId() === $user->getId()) {
-                    $this->addFlash(
-                        'error_admin',
-                        $this->translator->trans('administratorsCannotBanThemselves', [], 'controllers')
+            if (!$userUpdateDTO->editingAdmin) {
+                if ($userUpdateDTO->banned && $initialBannedAt === null) {
+                    $user->setBannedAt(new DateTime());
+                    $this->profileManager->disableProfiles(
+                        $user,
+                        UserRadiusProfileRevokeReason::ADMIN_BANNED_USER->value,
+                        true
                     );
-                    return $this->redirectToRoute('admin_user_edit', ['id' => $user->getId()]);
                 }
-                $user->setBannedAt(new DateTime());
-                $this->profileManager->disableProfiles(
-                    $user,
-                    UserRadiusProfileRevokeReason::ADMIN_BANNED_USER->value,
-                    true
-                );
-            } else {
-                $user->setBannedAt(null);
-                if ($form->get('isVerified')->getData()) {
+
+                if (!$userUpdateDTO->banned) {
+                    $user->setBannedAt(null);
+                }
+
+                if ($userUpdateDTO->isVerified) {
                     $this->profileManager->enableProfiles($user);
                 } else {
                     $this->profileManager->disableProfiles(
@@ -409,19 +403,18 @@ class UsersManagementController extends AbstractController
                 }
             }
 
-            $userRepository->save($user, true);
+            $this->userRepository->save($user, true);
 
-            $eventMetadata = [
-                'ip' => $request->getClientIp(),
-                'user_agent' => $request->headers->get('User-Agent'),
-                'edited' => $user->getUuid(),
-                'by' => $currentUser->getUuid(),
-            ];
             $this->eventActions->saveEvent(
                 $user,
                 AnalyticalEventType::USER_ACCOUNT_UPDATE_FROM_UI->value,
                 new DateTime(),
-                $eventMetadata
+                [
+                    'ip' => $request->getClientIp(),
+                    'user_agent' => $request->headers->get('User-Agent'),
+                    'edited' => $user->getUuid(),
+                    'by' => $currentUser->getUuid(),
+                ]
             );
 
             $uuid = $user->getUuid();
@@ -573,6 +566,16 @@ class UsersManagementController extends AbstractController
             return $this->redirect($lastPage);
         }
 
+        $lastConnectedProfile = $this->radiusProfileRepository->findUserLastConnection($user);
+
+        if ($lastConnectedProfile instanceof UserRadiusProfile) {
+            $lastStartConnection = $lastConnectedProfile->getLastConnectionStartAt();
+            $lastStopConnection = $lastConnectedProfile->getLastConnectionStopAt();
+        } else {
+            $lastStartConnection = null;
+            $lastStopConnection = null;
+        }
+
         return $this->render(
             'dashboard/actions/edit.html.twig',
             [
@@ -582,6 +585,9 @@ class UsersManagementController extends AbstractController
                 'data' => $data,
                 'current_user' => $currentUser,
                 'context' => FirewallType::DASHBOARD->value,
+                'userUpdateDTO' => $userUpdateDTO,
+                'lastStartConnection' => $lastStartConnection,
+                'lastStopConnection' => $lastStopConnection,
             ]
         );
     }