+++ CHANGELOG.md
@@ -12,6 +12,27 @@
   minimizing redundant data processing and database queries.
 - Renamed some enums to singular form to follow symfony guidelines for enum classes.
 
+# Release V1.9.0
+
+- **Fix bug 500** on `/dashboard/statistics/freeradius` in case the connection details are invalid and the portal can
+  reach
+  the server, added a new JSON error message for details.
+- **Fix bug 500** on `/metrics` in case the configuration env details are disabled or missing. Added a new JSON error
+  message for details.
+- Fix bug with returns with admin reset password, the problem was related with the previous firewall configuration
+  implementation where no context has returned.
+- Update `/dashboard/edit/{id}` to use DTO's and live components for validation.
+- New info icon about uuid explanation when admin is editing a user. Check the following page `/dashboard/edit/{id}` for more details.
+- New command for freeradius profile connection of each user, now the `UserRadiusProfile` entity saves the start/end
+  connection of the user profiles when the freeradius server gets a new request. (For later graphics generations and
+  user details)
+- Also, this new command is configurable on the page `dashboard/settings/schedule` because he is also cron based.
+- New bundle installed `composer require symfony/lock` required for the command next execution only start when the
+  current active ends.
+- NPM webpack-cli deprecations fixed to the latest stabled release (
+    - (https://www.npmjs.com/package/webpack-cli/v/5.1.4)) compatible with "@symfony/webpack-encore": "^5.1.0".
+- New validation on the Admin Authentication Methods page to check whether the project has all required certificates.
+
 # Release V1.8.1
 
 - Removed duplicated field relative to the user account verification, (Account Verification & User